feat: add support for fixtures in cypress
* Bump Cypress to 7.1.0
* Bump MongoDB to 3.6.6
* Split helpers for unit testing to import only database helpers for Cypress
