+++ .env.test
@@ -2,6 +2,9 @@
 APP_DB_URI="mongodb://root:password@127.0.0.1:27017"
 APP_DB_NAME="test"
 
+# smtp settings
+APP_SMTP_PORT=1025
+
 # session settings
 APP_SESSION_SECRET=localSecret
 
@@ -11,4 +14,4 @@ APP_STORAGE_PORT=9000
 APP_STORAGE_SSL=false
 APP_STORAGE_ACCESS_KEY="AKIAIOSFODNN7EXAMPLE"
 APP_STORAGE_SECRET_KEY="wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-APP_STORAGE_BUKCET_TEST=app-test
+APP_STORAGE_BUCKET=app-test
