+++ package.json
@@ -68,7 +68,7 @@
         "@types/minio": "^7.0.7",
         "@types/mjml": "^4.7.0",
         "@types/mjml-react": "^1.0.6",
-        "@types/mongodb": "^3.6.11",
+        "@types/mongodb": "^3.6.12",
         "@types/node": "^14.14.34",
         "@types/node-fetch": "^2.5.8",
         "@types/nodemailer": "^6.4.1",
@@ -89,7 +89,7 @@
         "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
         "copy-webpack-plugin": "^8.0.0",
         "css-loader": "^5.2.0",
-        "cypress": "^6.6.0",
+        "cypress": "^7.1.0",
         "dotenv": "^8.2.0",
         "dotenv-expand": "^5.1.0",
         "enquirer": "^2.3.6",
@@ -172,7 +172,7 @@
         "minio": "^7.0.18",
         "mjml": "^4.8.1",
         "mjml-react": "^1.2.0",
-        "mongodb": "^3.6.4",
+        "mongodb": "^3.6.6",
         "node-fetch": "^2.6.1",
         "nodemailer": "^6.5.0",
         "rate-limiter-flexible": "^2.2.1",
